# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PawSmart API — AI Pet Care + E-Commerce
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: petType
          schema: { type: string }
        - in: query
          name: brand
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [popular, price_asc, price_desc, newest] }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Single product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /pets:
    get:
      operationId: listPets
      tags: [pets]
      summary: Get the current user's pet profiles
      responses:
        "200":
          description: Pets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Pet" }
    post:
      operationId: createPet
      tags: [pets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PetInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pet" }
  /pets/{id}:
    put:
      operationId: updatePet
      tags: [pets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PetInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pet" }
    delete:
      operationId: deletePet
      tags: [pets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCartItemInput" }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /cart/items/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCartItemInput" }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
    delete:
      operationId: removeCartItem
      tags: [cart]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Current user's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Checkout - convert cart to order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderInput" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, brand, category, petType, price, stock, description, image, weight, expiryDate, popular, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        brand: { type: string }
        category: { type: string, description: "food | treats | accessories | grooming | toys" }
        petType: { type: string, description: "dog | cat | small_pet" }
        price: { type: number }
        stock: { type: integer }
        description: { type: string }
        image: { type: string }
        weight: { type: string, description: "e.g. 1kg, 5kg, 250g" }
        expiryDate: { type: string }
        popular: { type: boolean }
        breed:
          type: string
          nullable: true
        ageGroup:
          type: string
          nullable: true
          description: "puppy | adult | senior | kitten | all"
        healthCondition:
          type: string
          nullable: true
        createdAt: { type: string }
...
    Pet:
      type: object
      required: [id, name, type, breed, ageYears, weightKg, healthNotes, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, description: "dog | cat" }
        breed: { type: string }
        ageYears: { type: number }
        weightKg: { type: number }
        healthNotes: { type: string }
        createdAt: { type: string }
    PetInput:
      type: object
      required: [name, type, breed, ageYears, weightKg]
      properties:
        name: { type: string }
        type: { type: string }
        breed: { type: string }
        ageYears: { type: number }
        weightKg: { type: number }
        healthNotes: { type: string }
    CartItem:
      type: object
      required: [productId, quantity, product]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        product: { $ref: "#/components/schemas/Product" }
    Cart:
      type: object
      required: [items, subtotal, shipping, total, count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        subtotal: { type: number }
        shipping: { type: number }
        total: { type: number }
        count: { type: integer }
    AddCartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer, minimum: 1 }
    UpdateCartItemInput:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 0 }
...
    OrderItem:
      type: object
      required: [productId, name, brand, image, price, quantity]
      properties:
        productId: { type: integer }
        name: { type: string }
        brand: { type: string }
        image: { type: string }
        price: { type: number }
        quantity: { type: integer }
    Order:
      type: object
      required: [id, items, subtotal, shipping, total, status, customerName, address, phone, createdAt]
      properties:
        id: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        subtotal: { type: number }
        shipping: { type: number }
        total: { type: number }
        status: { type: string, enum: [pending, shipped, delivered, cancelled] }
        customerName: { type: string }
        address: { type: string }
        phone: { type: string }
        createdAt: { type: string }
    CreateOrderInput:
      type: object
      required: [customerName, address, phone]
      properties:
        customerName: { type: string }
        address: { type: string }
        phone: { type: string }
        notes: { type: string }
...