# lib/api-zod/src/generated/types/petInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PawSmart API — AI Pet Care + E-Commerce
 * OpenAPI spec version: 0.1.0
 */

export interface PetInput {
  name: string;
  type: string;
  breed: string;
  ageYears: number;
  weightKg: number;
  healthNotes?: string;
}